requirements.txt
absl-py==2.2.2
chex==0.1.89
etils==1.12.2
flax==0.10.6
fsspec==2025.3.2
glfw==2.9.0
humanize==4.12.3
importlib_resources==6.5.2
jax==0.6.0
jaxlib==0.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
ml_dtypes==0.5.1
msgpack==1.1.0
mujoco==3.3.2
mujoco-mjx==3.3.2
nest-asyncio==1.6.0
numpy==2.2.6
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.13
protobuf==6.31.0
Pygments==2.19.1
pynput==1.8.1
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
PyOpenGL==3.1.9
PyYAML==6.0.2
rich==14.0.0
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
tensorstore==0.1.75
toolz==1.0.0
treescope==0.1.9
trimesh==4.6.10
typing_extensions==4.13.2
zipp==3.21.0
